+++ client/src/App.js
@@ -6,6 +6,9 @@ import Landing from './components/layout/Landing';
 import Register from './components/auth/Register';
 import Login from './components/auth/Login';
 import Alert from './components/layout/Alert';
+import Dashboard from './components/dashboard/Dashboard';
+
+import PrivateRoute from './components/routing/PrivateRoute';
 
 import { UserContext } from './UserContext';
 import setAuthToken from './utils/setAuthToken';
@@ -39,6 +42,9 @@ const App = () => {
                         <Route exact path='/login'>
                             <Login />
                         </Route>
+                        <PrivateRoute exact path='/dashboard'>
+                            <Dashboard />
+                        </PrivateRoute>
                     </Switch>
                 </section>
             </>