+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
--
+-   `PrivateRoute` component
+-   Implement first frontend private route with /dashboard
+-   Dashboard UI
 
 ---